requirements.txt
ffmpeg==1.4
ffmpeg_python==0.2.0
imutils==0.5.4
ipython==8.9.0
librosa==0.9.2
matplotlib==3.6.2
mediapipe==0.9.0.1
more_itertools==9.0.0
moviepy==1.0.3
nltk==3.8.1
noisereduce==2.0.1
numpy==1.23.4
opencv_python==4.5.5.62
pandas==1.5.1
Pillow==9.4.0
psutil==5.9.4
pytest==7.2.1
pytube==12.1.2
regex==2022.10.31
scikit_learn==1.2.1
scipy==1.10.0
setuptools==65.5.0
spacy==3.5.0
# torch==1.13.1+cu116  # Kept as comment for version reference
tqdm==4.64.1
transformers==4.25.1
python-vlc==3.0.18121
